refactor(wallet_policy): isolate policy invariants
- keep validated state construction and mutation in the private policy module
- move translation and public API implementations to the parent module
